Cache balance and transactions in mongodb

+++ backend/src/model/balance_response.rs
@@ -4,5 +4,5 @@ use serde::Serialize;
 #[derive(Debug, Serialize)]
 pub struct BalanceResponse {
     pub hledger: Decimal,
-    pub real: Decimal,
+    pub real: f64,
 }
